confirm button to delete

+++ public/schedule/prepare.js
@@ -5,27 +5,32 @@ const urlContent = document.querySelector('#urlContent');
 
 
 function deleteAllNotifications(){
-  deleteAllNotificationsBtn.disabled = true;
-  fetch('/deleteprojectnotifications', {
-    method:'POST',
-    headers: {
-      'Content-Type':'application/json',
-      'Accept':'application/json',
-    },
-    body: JSON.stringify({
-      name: 'notification',
+  const deleteAll = confirm("Are you sure you want to delete all notifications? The action is not reversible.");
+  if (deleteAll === true) {
+    deleteAllNotificationsBtn.disabled = true;
+    fetch('/deleteprojectnotifications', {
+      method:'POST',
+      headers: {
+        'Content-Type':'application/json',
+        'Accept':'application/json',
+      },
+      body: JSON.stringify({
+        name: 'notification',
+      })
     })
-  })
-    .then(res => {
-      if(res.url && res.ok){
-        window.location = res.url;
-      }
-      deleteAllNotificationsBtn.disabled = false;
-    })
-    .catch(err => {
-      console.log(err);
-      deleteAllNotificationsBtn.disabled = false;
-    })
+      .then(res => {
+        if(res.url && res.ok){
+          window.location = res.url;
+        }
+        deleteAllNotificationsBtn.disabled = false;
+      })
+      .catch(err => {
+        console.log(err);
+        deleteAllNotificationsBtn.disabled = false;
+      })
+  } else {
+    // console.log('Cancelled');
+  };
 };
 
 function removeScheduleDateTimePicker(){
